requirements.txt
# Core
python-telegram-bot==21.7
openai>=1.80.0
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.40
asyncpg>=0.30.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.10.0
httpx>=0.28.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0

# Optional
redis>=5.2.0
pgvector>=0.3.0
